refactor: promise utils
