+++ apps/main/src/dex/store/createDashboardSlice.ts
@@ -17,10 +17,11 @@ import { ChainId, claimButtonsKey, CurveApi, FnStepResponse, PoolDataMapper } fr
 import { fulfilledValue, getStorageValue, setStorageValue } from '@/dex/utils'
 import type { IProfit } from '@curvefi/api/lib/interfaces'
 import { useWallet } from '@evm-ui/features/connect-wallet'
-import { Chain, getErrorMessage, sleep } from '@evm-ui/utils'
+import { Chain, getErrorMessage } from '@evm-ui/utils'
 import { setMissingProvider } from '@evm-ui/utils/store.util'
 import { shortenAccount } from '@legacy-ui/utils'
 import type { Address } from '@primitives/address.utils'
+import { sleep } from '@primitives/promise.utils'
 import { PromisePool } from '@supercharge/promise-pool'
 import { userPoolRewardCrvApy } from '../queries/user-pool-reward-crv-apy.query'
 import { fetchUserPools } from '../queries/user-pools.query'
