+++ packages/evm-ui/src/utils/time.ts
@@ -51,29 +51,6 @@ export const TIME_OPTION_MS = {
   '1Y': toMilliseconds(SECONDS_PER_YEAR),
 } as const
 
-export const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms))
-
-/** Handles a promise with a timeout. If the promise does not resolve within the specified timeout, it rejects with an error. */
-const handleTimeout = <T>(promise: Promise<T>, timeout: number, message?: string): Promise<T> =>
-  new Promise((resolve, reject) => {
-    const id = setTimeout(() => {
-      clearTimeout(id)
-      reject(new Error(message || `Promise timed out after ${timeout}ms`))
-    }, timeout)
-    promise.then(resolve, reject).finally(() => clearTimeout(id))
-  })
-
-/** Waits for a callback to return a truthy value, polling at specified intervals, with a timeout. */
-export const waitFor = async (
-  callback: () => unknown,
-  { timeout, step = 1000, message }: { timeout: number; message?: string; step?: number },
-) => {
-  const waitUntil = async () => {
-    while (!(await callback())) await sleep(step)
-  }
-  await handleTimeout<void>(waitUntil(), timeout, message)
-}
-
 export const formatTimeDiff = (start: Date) => `${(new Date().getTime() - start.getTime()).toLocaleString()}ms`
 
 const TIME_UNITS = [
