+++ tests/cypress/component/llamalend/loan.rpc.cy.tsx
@@ -32,9 +32,9 @@ import { fundErc20, fundEth } from '@cy/support/helpers/tenderly/vnet-fund'
 import { LOAD_TIMEOUT, skipTestsAfterFailure } from '@cy/support/ui'
 import { getLib } from '@evm-ui/features/connect-wallet'
 import { MarketType } from '@evm-ui/types/market'
-import { waitFor } from '@evm-ui/utils'
 import type { Decimal } from '@primitives/decimal.utils'
 import { recordValues } from '@primitives/objects.utils'
+import { waitFor } from '@primitives/promise.utils'
 
 const testCases = recordValues(MarketType).map(marketType => oneLoanTestMarket(marketType))
 /**
